Update autocomplete.rkt
Hi, sorry if this is a waste of time for you, I was just thinking of adding a couple more shortcuts that I think some students might find helpful especially with where we are in the course right now. I am a CS135 student right now and this is honestly such a useful program which is why I'm trying to add on to it. Please don't pay it any mind if you don't like these changes I won't be bothered.